// src/state/types/Setting.ts
export enum Setting {
  mapType = 'mapType',
  isMapLocked = 'isMapLocked',
  isGeoLocEnabled = 'isGeoLocEnabled', // TODO: not actually used atm
  mapRegion = 'mapRegion',
  locUpdateInterval = 'locUpdateInterval',
  locUpdateDistance = 'locUpdateDistance',
  lineWidth = 'lineWidth',
  lineColor = 'lineColor',
}